Changes Clock properties from Int to Long
To match the generally used type for milliseconds.

+++ app/src/main/java/alexman/chessclock_multiplayer/model/Clock.kt
@@ -4,23 +4,23 @@ package alexman.chessclock_multiplayer.model
 data class Clock private constructor( // use new() and load() instead of constructor
     val profile: Profile,
     val timeControl: TimeControl,
-    val timeLeftMillis: Int = timeControl.timeSeconds * 1000,
-    val lastSessionTimeMillis: Int = timeLeftMillis,
+    val timeLeftMillis: Long = timeControl.timeSeconds * 1000L,
+    val lastSessionTimeMillis: Long = timeLeftMillis,
 ) {
     companion object {
         fun new(profile: Profile, timeControl: TimeControl): Clock =
             Clock(
                 profile,
                 timeControl,
-                timeLeftMillis = timeControl.timeSeconds * 1000,
-                lastSessionTimeMillis = timeControl.timeSeconds * 1000,
+                timeLeftMillis = timeControl.timeSeconds * 1000L,
+                lastSessionTimeMillis = timeControl.timeSeconds * 1000L,
             )
 
         fun load(
             profile: Profile,
             timeControl: TimeControl,
-            timeLeftMillis: Int,
-            lastSessionTimeMillis: Int
+            timeLeftMillis: Long,
+            lastSessionTimeMillis: Long,
         ): Clock = Clock(profile, timeControl, timeLeftMillis, lastSessionTimeMillis)
     }
 }
